feat!: simplify Pointer and Token
Re-reading the spec, I found that `-` is a valid `Token` at any position
in a `Pointer`. This is simpler to handle overall, which I take
advantage of now.

It just so happens that the semantics of `-` make it never useful
anywhere but in the last position (to denote an append operation),
because it'll always fail otherwise. The new lenses for `Pointer` work
this way, again making the usage within `Apply` more uniform.