feat!: move away from attoparsec for parsing Token
This was more complicated and didn't work.

The fact that anything not all digits should become `K` but, when it is
all digits, it cannot contain leading zeros, meant that we had two
failure cases within the parser for numbers: one that should trip over
to `K` (via `<|>`) and one that should fail the entire parser (leading
zeros). I was unable to solve this. There is also the case of something
like `"1e0"` where the `"1"` would parse as `N "1"` and leave `"eO"`
unconsumed. I couldn't make this work either. Add to all that the fact
that tests use negative indexes (even though the spec doesn't allow
that), and it all became too much. Some straight `Text`-reading seems
clear enough now.

I am still using `attoparsec` for the `Pointer`, which seems a better
use-case due to the way `sepBy` behaves compared to a naive `splitOn
'/'` w.r.t. duplicate separators and other things. The dependency is not
worth removing because anyone using this library is depending on it
through `aeson` anyway.